GeneradorCertificados/requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.9
idna==3.3
mysqlclient==2.0.3
numpy==1.21.4
pandas==1.3.4
Pillow==9.1.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
